+++ server/src/main/java/org/allaymc/server/blockentity/component/BlockEntityChestContainerHolderComponentImpl.java
@@ -1,9 +1,10 @@
 package org.allaymc.server.blockentity.component;
 
-import lombok.Getter;
 import org.allaymc.api.block.data.BlockFace;
-import org.allaymc.api.blockentity.component.BlockEntityContainerHolderComponent;
+import org.allaymc.api.blockentity.component.BlockEntityChestContainerHolderComponent;
 import org.allaymc.api.blockentity.component.BlockEntityPairableComponent;
+import org.allaymc.api.blockentity.interfaces.BlockEntityChest;
+import org.allaymc.api.container.Container;
 import org.allaymc.api.container.interfaces.BlockContainer;
 import org.allaymc.api.eventbus.EventHandler;
 import org.allaymc.api.math.position.Position3i;
@@ -16,8 +17,8 @@ import org.allaymc.server.container.impl.DoubleChestContainerImpl;
 /**
  * @author IWareQ | daoge_cmd
  */
-public class BlockEntityChestContainerHolderComponentImpl extends BlockEntityContainerHolderComponentImpl {
-    @Getter
+public class BlockEntityChestContainerHolderComponentImpl extends BlockEntityContainerHolderComponentImpl implements BlockEntityChestContainerHolderComponent {
+
     protected final DoubleChestContainerImpl doubleChestContainer;
 
     @Dependency
@@ -26,6 +27,36 @@ public class BlockEntityChestContainerHolderComponentImpl extends BlockEntityCon
     public BlockEntityChestContainerHolderComponentImpl() {
         super(ChestContainerImpl::new);
         this.doubleChestContainer = new DoubleChestContainerImpl();
+    }
+
+    @Override
+    public Container getDoubleChestContainer() {
+        if (this.pairableComponent.isPaired()) {
+            // Update the double chest container object if this chest is paired
+            var blockPos = this.baseComponent.getPosition();
+            var pairChest = (BlockEntityChest) this.pairableComponent.getPair();
+            var left = getContainer();
+            var right = pairChest.getContainer();
+            if (!this.pairableComponent.isLead()) {
+                // Block pos is always the pos of the lead chest
+                blockPos = pairChest.getPosition();
+                // Swap left and right container if the current chest is not the lead chest
+                var temp = left;
+                left = right;
+                right = temp;
+            }
+
+            this.doubleChestContainer.setLeft(left);
+            this.doubleChestContainer.setRight(right);
+            this.doubleChestContainer.setBlockPos(blockPos);
+        }
+
+        return this.doubleChestContainer;
+    }
+
+    // Used in BlockEntityChestBaseComponentImpl
+    public Container getDoubleChestContainerDirectly() {
+        return this.doubleChestContainer;
     }
 
     @EventHandler
@@ -43,23 +74,12 @@ public class BlockEntityChestContainerHolderComponentImpl extends BlockEntityCon
         }
 
         if (pairableComponent.isPaired()) {
-            var pair = pairableComponent.getPair();
-            if (!hasSpaceAbove(pair.getPosition())) {
+            if (!hasSpaceAbove(pairableComponent.getPair().getPosition())) {
                 return;
             }
 
-            var left = getContainer();
-            var right = ((BlockEntityContainerHolderComponent) pairableComponent.getPair()).getContainer();
-            if (!pairableComponent.isLead()) {
-                var temp = left;
-                left = right;
-                right = temp;
-            }
-
-            this.doubleChestContainer.setLeft(left);
-            this.doubleChestContainer.setRight(right);
-            this.doubleChestContainer.setBlockPos(new Position3i(interactInfo.clickedBlockPos(), event.getDimension()));
-            this.doubleChestContainer.addViewer(player.getController());
+            // Use getDoubleChestContainer() method here which will update the double chest container object
+            getDoubleChestContainer().addViewer(player.getController());
         } else {
             if (this.container instanceof BlockContainer blockContainer) {
                 blockContainer.setBlockPos(this.baseComponent.getPosition());