+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Unless otherwise specified, any version comparison below is the comparison of th
 - (API) Added method `Player.transfer()` to transfer the player to another server.
 - (API) Added method `Entity.setImmobile()` to make the entity immobile.
 - (API) Added method `Container.isFull()` to check if a container is full.
+- (API) Added method `BlockEntityChest.getDoubleChestContainer()` to get the container object to use when the chest is paired.
 - Added properties `network-settings.raknet-send-cookie` and `network-settings.raknet-max-mtu` to `server.properties`.
 - Implemented hopper item transfer with furnace and brewing stand slot rules.
 - Implemented Totem of Undying activation for players.